use ArticleVersionDto in ArticleDataHandler, update some namespaces, rename and move SalesforceErrorDto
