+++ App.Salesforce.Cms/Actions/ArticleActions.cs
@@ -1,5 +1,4 @@
-﻿using App.Salesforce.Cms.Models.Dtos;
-using App.Salesforce.Cms.Models.Requests;
+﻿using App.Salesforce.Cms.Models.Requests;
 using App.Salesforce.Cms.Models.Responses;
 using Apps.Salesforce.Cms;
 using Apps.Salesforce.Cms.Api;
