+++ App.Salesforce.Cms/Models/Dtos/AdditionalInformationDto.cs
@@ -1,10 +1,10 @@
 ﻿using Blackbird.Applications.Sdk.Common;
 
-namespace App.Salesforce.Cms.Models.Dtos;
+namespace Apps.Salesforce.Cms.Models.Dtos;
 
 public class AdditionalInformationDto
 {
-    [Display("Can Aarchive")]
+    [Display("Can Archive")]
     public bool CanArchive { get; set; }
 
     [Display("Can delete")]
