+++ App.Salesforce.Cms/Models/Dtos/DraftErrorDto.cs
@@ -1,4 +1,4 @@
-namespace App.Salesforce.Cms.Models.Dtos;
+namespace Apps.Salesforce.Cms.Models.Dtos;
 
 public class DraftErrorDto
 {
