+++ App.Salesforce.Cms/Polling/ArticlePollingList.cs
@@ -1,5 +1,4 @@
-﻿using App.Salesforce.Cms.Models.Dtos;
-using App.Salesforce.Cms.Models.Responses;
+﻿using App.Salesforce.Cms.Models.Responses;
 using Apps.Salesforce.Cms.Api;
 using Apps.Salesforce.Cms.Models.Dtos;
 using Apps.Salesforce.Cms.Models.Requests;
